+++ R/shoji-order-slots.R
@@ -118,6 +118,10 @@ setMethod("duplicates<-", c("ShojiOrder", "logical"), function (x, value) {
     x@duplicates <- value
     grps <- vapply(x@graph, inherits, logical(1), what="OrderGroup")
     x@graph[grps] <- lapply(x@graph[grps], `duplicates<-`, value=value)
+    if (!value) {
+        ## We're setting duplicates: FALSE, so dedupe
+        x <- dedupeOrder(x)
+    }
     return(x)
 })
 #' @rdname ShojiOrder-slots
@@ -127,6 +131,9 @@ setMethod("duplicates<-", c("OrderGroup", "logical"), function (x, value) {
     x@duplicates <- value
     grps <- vapply(x@entities, inherits, logical(1), what="OrderGroup")
     x@entities[grps] <- lapply(x@entities[grps], `duplicates<-`, value=value)
+    ## Note: not calling dedupeOrder here because it's most likely that this is
+    ## only called from within the duplicates<- method for ShojiOrder, which
+    ## does the deduping
     return(x)
 })
 #' @rdname ShojiOrder-slots